Cleaning up logs and functions
