+++ src/schema/CombinedSchemas.ts
@@ -1,9 +1,11 @@
+import { LoggerFactory } from '../telemetry/LoggerFactory';
 import { PrivateSchemas, PrivateSchemasType } from './PrivateSchemas';
 import { RegionalSchemas, RegionalSchemasType } from './RegionalSchemas';
 import { ResourceSchema } from './ResourceSchema';
 import { SamSchemas, SamSchemasType } from './SamSchemas';
 
 export class CombinedSchemas {
+    private static readonly log = LoggerFactory.getLogger('CombinedSchemas');
     readonly numSchemas: number;
     readonly schemas: Map<string, ResourceSchema>;
 
@@ -16,13 +18,6 @@ export class CombinedSchemas {
         this.numSchemas = this.schemas.size;
     }
 
-    toLog() {
-        return {
-            schema: this.schemas.size,
-            names: [...this.schemas.keys()],
-        };
-    }
-
     static from(
         regionalSchemas?: RegionalSchemasType,
         privateSchemas?: PrivateSchemasType,
@@ -32,6 +27,9 @@ export class CombinedSchemas {
         const privateSchema = privateSchemas === undefined ? undefined : PrivateSchemas.from(privateSchemas);
         const samSchema = samSchemas === undefined ? undefined : SamSchemas.from(samSchemas);
 
+        CombinedSchemas.log.info(
+            `Schemas from ${regionalSchemas?.schemas.length} public schemas, ${privateSchema?.schemas.size} private schemas and ${samSchema?.schemas.size} sam schemas`,
+        );
         return new CombinedSchemas(regionalSchema, privateSchema, samSchema);
     }
 }
